requirements.txt
openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.21.0
sounddevice>=0.4.5
soundfile>=0.12.1
pyaudio>=0.2.11
pydub>=0.25.1
tkinter-tooltip>=2.1.0
keyboard>=0.13.5
pyperclip>=1.8.2
scipy>=1.9.0
librosa>=0.10.0
noisereduce>=3.0.0
customtkinter>=5.2.0
Pillow>=9.0.0
threading-timer>=0.1.0
requests>=2.28.0 